Allow for the deletion of classes unless they have decks (next step)

+++ app/api/class_routes.py
@@ -1,5 +1,5 @@
-from flask import Blueprint, jsonify, session, request
-from app.models import User, db, Class, CategoryClass
+from flask import Blueprint, Response, jsonify, session, request
+from app.models import User, db, Class, CategoryClass, Deck
 from flask_login import current_user, login_user, logout_user, login_required
 
 class_routes = Blueprint('class', __name__)
@@ -9,7 +9,8 @@ class_routes = Blueprint('class', __name__)
 def get_all_classes():
     if current_user.is_authenticated:
     
-        classes = Class.query.all()
+        user_id = current_user.id
+        classes = Class.query.filter(Class.user_id == user_id).all()
     
     all_classes = [ class_var.to_dict() for class_var in classes ] # class is a reserved keyword
     
@@ -41,19 +42,25 @@ def create_classes():
 ## delete a class
 @class_routes.route('/<int:id>', methods = ["DELETE"])
 def delete_class_by_id(id):
+    
+    
     if current_user.is_authenticated:
         class_obj = Class.query.get(id)
         
         if not class_obj: 
-            return "Class not found", 401 # 401 is unathorized
+            return {"message": "Class not found"}, 401 # 401 is unathorized
 
         if class_obj.user_id != current_user.id:
-            return "User does not own class", 404
+            return {"message": "User does not own class"}, 404
+        
+        list_of_decks = Deck.query.filter(Deck.class_id == class_obj.id).all()
+        
+        if len(list_of_decks) > 0:
+            return {"message": "All decks must be deleted before a class is deleted"}, 409
         
         # Delete the category class association when the class is deleted    
         try:
             category_classes = CategoryClass.query.filter(CategoryClass.class_id == id).all()
-            print(category_classes)
             cat_class = category_classes[0]
             db.session.delete(cat_class)
             db.session.commit()
@@ -64,6 +71,6 @@ def delete_class_by_id(id):
             db.session.delete(class_obj)
             db.session.commit()
         except:
-            return "Class cannot be deleted"
+            return {"message": "Class cannot be deleted"}, 500
         
-    return "Class successfully deleted"
+    return {"message": "Class successfully deleted"}
